Renamed: pathmanager and Fixed
