+++ stobu/core/rubibuilder.py
@@ -6,7 +6,7 @@ import re
 
 
 # My Modules
-from stobu import projectpathmanager as ppath
+from stobu.tools import pathmanager as ppath
 from stobu.util import assertion
 from stobu.util.fileio import read_file_as_yaml
