+++ stobu/core/storydatacreator.py
@@ -7,8 +7,8 @@
 # My Modules
 from stobu.dataconverter import conv_to_story_record
 from stobu.datatypes import StoryData
+from stobu.tools import pathmanager as ppath
 from stobu.util import assertion
-from stobu import projectpathmanager as ppath
 from stobu.util.fileio import read_file_as_markdown, read_file_as_yaml
 from stobu.util.log import logger
