+++ stobu/tools/pathmanager.py
